Fixing design page issues
