+++ src/assets/styles/main.css
@@ -1377,7 +1377,7 @@ h2 {
 }
 
 .button-1 {
-    margin-top: 180px;
+    margin-left: 150px;
     height: 30px;
     width: 200px;
     font-size: 10px;
