+++ src/routes/pathnames.ts
@@ -4,9 +4,7 @@ export const authRoute = '/auth';
 export const pwdRoute = '/forgot';
 export const signupRoute = '/signup'
 export const settingsRoute = '/settings'
-export const createAccountRoute = '/create';
-export const verifyOrgRoute = '/verify';
-export const processingTextRoute = '/confirm';
+
 // Private routes
 export const providerRoute = '/provider';
 export const formRoute = '/provider/add';
